separate extension code into different files for readability
